Order-agnostic Binary Search
Given a sorted array of numbers, find if a given number ‘key’ is present
in the array. Though we know that the array is sorted, we don’t know if
it’s sorted in ascending or descending order. You should assume that the
array can have duplicates.

Write a function to return the index of the ‘key’ if it is present in
the array, otherwise return -1.

Example 1: